cdk/requirements.txt
aws-cdk-lib==2.44.0
constructs>=10.0.0,<11.0.0
oe-patterns-cdk-common@git+https://github.com/ordinaryexperts/aws-marketplace-oe-patterns-cdk-common@11c1ecd14957d01a6e212efd7f385e997cde2c02
